fix: some

+++ microservices/compress/compress/delivery/compress_microservice.go
@@ -17,13 +17,16 @@ type CompressManager struct {
 func NewCompressManager(compressUsecase compressinterfaces.ICompressUsecase, logger *logrus.Logger) *CompressManager {
 	return &CompressManager{
 		compressUsecase: compressUsecase,
-		logger: &logrus.Entry{Logger: logger},
+		logger:          &logrus.Entry{Logger: logger},
 	}
 }
 
 func (cm *CompressManager) CompressAndSaveFile(ctx context.Context, in *compress.CompressAndSaveFileInput) (*compress.Nothing, error) {
 	funcName := "CompressDelivery.CompressAndSaveFile"
 	cm.logger.Debugf("%s: got request: %s", funcName, in)
+	if in == nil {
+		return &compress.Nothing{}, compressinterfaces.WrongData
+	}
 	err := cm.compressUsecase.CompressAndSaveFile(in.FileName, in.FileType, in.File)
 	return &compress.Nothing{}, err
 }
@@ -31,6 +34,9 @@ func (cm *CompressManager) CompressAndSaveFile(ctx context.Context, in *compress
 func (cm *CompressManager) DeleteFile(ctx context.Context, in *compress.DeleteFileInput) (*compress.Nothing, error) {
 	funcName := "CompressDelivery.DeleteFile"
 	cm.logger.Debugf("%s: got request: %s", funcName, in)
+	if in == nil {
+		return &compress.Nothing{}, compressinterfaces.WrongData
+	}
 	err := cm.compressUsecase.DeleteFile(in.FileName)
 	return &compress.Nothing{}, err
 }
